requirements.txt
apache-airflow[celery]==2.10.0
--constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.10.0/constraints-3.12.txt
apache-airflow-providers-snowflake
python-dotenv
sqlparse
jupyterlab
snowflake-connector-python
openpyxl
xlrd